moved logic into separate unit tests.
